:recycle: 画面の大きさを決定する部分を抜き出す
