+++ src/hooks/useIsMobile.ts
@@ -1,0 +1,17 @@
+import { useEffect, useState } from "react";
+
+export const useIsMobile = (): boolean => {
+    const [isMobile, setIsMobile] = useState(false);
+
+    useEffect(() => {
+        const handleResize = () => {
+            setIsMobile(window.innerWidth < 1024);
+        };
+
+        handleResize();
+        window.addEventListener("resize", handleResize);
+        return () => window.removeEventListener("resize", handleResize);
+    }, []);
+
+    return isMobile;
+};
